move functions to utils, address comments
